Implemented (back-)references to character ranges.
In this initial implementation, the same struct's are used for ranges
and references.  This results in small code size, but it wastes the
range struct's, which are pretty large, and it has some obscure properties
(e.g., references are counted the same as ranges for purposes of numbering
by even further references, which is inconsistent with sed back-references).

CVS-ID: src/rpp.c 1.3