Changed another wordlist mode rule to avoid producing duplicates, which would be produced in some special cases.  Unfortunately, this same change increases processing cost of the rule itself.  Additionally, because of the discrepancy in minimum length requirement with the "lowercase" rule found earlier in the ruleset, this change may remove some candidate passwords that would be generated prior to the change.  Arguably, this is acceptable because the minimum length requirements were/are arbitrary anyway and because the previous behavior was not implemented on purpose.
Specifically, if the input wordlist contains "Bed" (a three-letter word,
capitalized, ending in "ed"), then previously the word "bed" would be
generated.  With this change, it would not be generated anymore because
"Bed" does not pass the requirement for the lowercase rule (must be four
alphanumeric characters or longer), yet the changed rule with "P" will
no longer act as lowercase-only for words already ending in "ed" (it will
discard such words).

CVS-ID: run/john.conf 1.38